backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pydantic==2.5.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
diffusers==0.24.0
transformers==4.36.0
torch>=2.0.0
torchvision
accelerate
controlnet-aux
xformers
opencv-python==4.8.1.78
numpy==1.24.3
aiofiles==23.2.1
httpx==0.25.2
huggingface_hub>=0.19.0
stripe>=8.0.0
